+++ mep/common/views.py
@@ -1,7 +1,11 @@
+import calendar
+
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.http import Http404, JsonResponse
-from django.utils.cache import patch_vary_headers
+from django.utils.cache import get_conditional_response, patch_vary_headers
 from django.views.generic.base import ContextMixin, TemplateResponseMixin, View
+from parasolr.django.queryset import SolrQuerySet
+from parasolr.utils import solr_timestamp_to_datetime
 import rdflib
 
 from mep.common import SCHEMA_ORG
@@ -13,7 +17,7 @@ class LoginRequiredOr404Mixin(LoginRequiredMixin):
     to the login form.'''
 
     def handle_no_permission(self):
-        # if permission is denied, raise a 404
+        '''If permission is denied, raise an :class:`~django.http.Http404`'''
         raise Http404
 
 
@@ -51,12 +55,14 @@ class LabeledPagesMixin(ContextMixin):
         return context
 
     def dispatch(self, request, *args, **kwargs):
+        '''Wrap the dispatch method to patch in page label header for
+        ajax requests.'''
         response = super(LabeledPagesMixin, self).dispatch(request, *args, **kwargs)
-        if self.request.is_ajax():
         # NOTE we need to replace the en dashes in alpha pagelabels with a
         # hyphen when requested via ajax because unicode can't be sent via the
         # X-Page-Labels header. this needs to get converted back to an en dash
         # on the client side.
+        if self.request.is_ajax():
             response['X-Page-Labels'] = '|'.join(
                 [label.replace('–', '-') for index, label in self._page_labels])
         return response
@@ -161,6 +167,7 @@ class AjaxTemplateMixin(TemplateResponseMixin, VaryOnHeadersMixin):
         return super().get_template_names()
 
     def dispatch(self, request, *args, **kwargs):
+        '''Set a total result header on the response'''
         response = super(AjaxTemplateMixin, self).dispatch(request, *args, **kwargs)
         response['X-Total-Results'] = self.get_queryset().count()
         return response
@@ -185,3 +192,56 @@ class FacetJSONMixin(TemplateResponseMixin, VaryOnHeadersMixin):
         '''Construct a JsonResponse based on the already-populated queryset
         data for the view.'''
         return JsonResponse(self.object_list.get_facets())
+
+
+# last modified view mixin adapted from ppa
+
+
+class SolrLastModifiedMixin(View):
+    """View mixin to add last modified headers based on Solr"""
+
+    #: solr query filter for getting last modified date
+    solr_lastmodified_filters = {}   # by default, find all
+
+    def get_solr_lastmodified_filters(self):
+        '''Get filters for last modified Solr query. By default returns
+        :attr:`solr_lastmodified_filters`.'''
+        return self.solr_lastmodified_filters
+
+    def last_modified(self):
+        '''Return last modified :class:`datetime.datetime` from the
+        specified Solr query'''
+        filter_qs = self.get_solr_lastmodified_filters()
+        sqs = SolrQuerySet().filter(**filter_qs) \
+            .order_by('-last_modified').only('last_modified')
+        try:
+            # Solr stores date in isoformat; convert to datetime
+            return solr_timestamp_to_datetime(sqs[0]['last_modified'])
+            # skip extra call to Solr to check count and just grab the first
+            # item if it exists
+        except (IndexError, KeyError):
+            # if a syntax or other solr error happens, no date to return
+            pass
+
+    def dispatch(self, request, *args, **kwargs):
+        '''Wrap the dispatch method to add a last modified header if
+        one is available, then return a conditional response.'''
+
+        # NOTE: this doesn't actually skip view processing,
+        # but without it we could return a not modified for a non-200 response
+        response = super(SolrLastModifiedMixin, self) \
+            .dispatch(request, *args, **kwargs)
+
+        last_modified = self.last_modified()
+        if last_modified:
+            # remove microseconds so that comparison will pass,
+            # since microseconds are not included in the last-modified header
+            last_modified = last_modified.replace(microsecond=0)
+            response['Last-Modified'] = last_modified \
+                .strftime('%a, %d %b %Y %H:%M:%S GMT')
+            # convert the same way django does so that they will
+            # compare correctly
+            last_modified = calendar.timegm(last_modified.utctimetuple())
+
+        return get_conditional_response(request, last_modified=last_modified,
+                                        response=response)