+++ srcmedia/ts/lib/common.ts
@@ -163,6 +163,19 @@ function pluralize (n: number|string|Array<any>) {
 }
 
 /**
+ * Adds a suffix to a number to convert it to ordinal form (1st, 3rd, etc.)
+ *
+ * @param n number to be converted
+ */
+function ordinalize (n: number) {
+    const j = n % 10, k = n % 100
+    if (j == 1 && k != 11) return n + "st"
+    if (j == 2 && k != 12) return n + "nd"
+    if (j == 3 && k != 13) return n + "rd"
+    return n + "th"
+}
+
+/**
  * Set the height of a set of sibling elements by traversing sibling pointers
  * recursively until all siblings have been reached.
  *
@@ -174,6 +187,18 @@ function setAllHeights($element: HTMLElement, height: string) {
     if ($element.nextElementSibling) {
         setAllHeights(($element.nextElementSibling as HTMLElement), height)
     }
+}
+
+/**
+ * Reverse a slugify-like operation, decoding a string like one_two_three to one
+ * like "One Two Three".
+ *
+ * @param str string to operate on
+ */
+function unslugify(str: string) {
+    return str.split('_')
+              .map(part => part.charAt(0).toUpperCase() + part.substr(1))
+              .join(' ')
 }
 
 export {
@@ -189,5 +214,7 @@ export {
     fakeToggle,
     toggleTab,
     pluralize,
-    setAllHeights
+    ordinalize,
+    setAllHeights,
+    unslugify
 }